+++ src/collective/lazysizes/setuphandlers.py
@@ -1,13 +1,23 @@
 # -*- coding: utf-8 -*-
 from Products.CMFPlone.interfaces import INonInstallable
+from Products.CMFQuickInstallerTool import interfaces as BBB
 from zope.interface import implementer
 
 
+@implementer(BBB.INonInstallable)  # BBB: Plone 4.3
 @implementer(INonInstallable)
-class HiddenProfiles(object):
+class NonInstallable(object):  # pragma: no cover
+
+    @staticmethod
+    def getNonInstallableProducts():
+        """Hide in the add-ons configlet."""
+        return [
+            u'collective.lazysizes.upgrades.v10',
+        ]
 
     def getNonInstallableProfiles(self):  # pragma: no cover
-        """Do not show on Plone's list of installable profiles."""
+        """Hide at site creation."""
         return [
+            u'collective.lazysizes.upgrades.v10:default',
             u'collective.lazysizes:uninstall',
         ]